requirements.txt
python-dotenv
langchain
langchain-chroma
langchain-gigachat
pypdf
python-telegram-bot
nest-asyncio
langchain_community
python-docx
pandas
openpyxl
xlrd
langchain-text-splitters
langchain-core
langchain-community